+++ framework/Furion.Pure/Localization/Options/LocalizationSettingsOptions.cs
@@ -9,6 +9,7 @@
 using Furion.ConfigurableOptions;
 using Microsoft.Extensions.Configuration;
 using System;
+using System.Reflection;
 
 namespace Furion.Localization
 {
@@ -33,6 +34,16 @@ namespace Furion.Localization
         public string DefaultCulture { get; set; }
 
         /// <summary>
+        /// 资源文件名前缀
+        /// </summary>
+        public string LanguageFilePrefix { get; set; }
+
+        /// <summary>
+        /// 资源所在程序集名称
+        /// </summary>
+        public string AssemblyName { get; set; }
+
+        /// <summary>
         /// 选项后期配置
         /// </summary>
         /// <param name="options"></param>
@@ -42,6 +53,8 @@ namespace Furion.Localization
             ResourcesPath ??= "Resources";
             SupportedCultures ??= Array.Empty<string>();
             DefaultCulture ??= string.Empty;
+            LanguageFilePrefix ??= "Lang";
+            AssemblyName ??= Assembly.GetEntryAssembly().GetName().Name;
         }
     }
 }